refactor: move TableUpdate & UpdateType to table types
